+++ R/bootSE.R
@@ -65,7 +65,7 @@
 #'     verbose = TRUE)
 #'
 #' fit.boot <- bootSE(fit, 50, use.mle = TRUE, control = list(
-#'     earlyPhase = 25, convCrit = "either",
+#'     burnin = 25, convCrit = "either",
 #'     tol0 = 6e-03, tol2 = 6e-03, mcmaxIter = 60))
 #' }
 bootSE <- function(object, nboot = 100, ci = 0.95, use.mle = TRUE,